refactor: 모달 열고 닫기 책임 분리
